Sketch
We intend to use IR sensors to monitor traffic through the doors. This can be accomplished by using two sensors to sense whether people are entering or leaving the building. Arduinos will run these sensors, and will write data to a file. The visualizations will read the file and create a visualization. In addition, musical notes will play everytime someone walks into the building. The visualization could be a line plot, or have a dot for every person in the building.